[MOB-24631] Add ability to rescan folders with templates without restarting
